+++ dotnet-uptime/Config/AppConfig.cs
@@ -1,4 +1,6 @@
 
+using Microsoft.Extensions.Logging;
+
 namespace MV10.DotnetUptime;
 
 /// <summary>
@@ -11,4 +13,5 @@ public class AppConfig
     public int MaxHistograms { get; set; } = 10;
     public int MaxTimeSeries { get; set; } = 1000;
     public bool ExcludeSelf { get; set; } = true;
+    public LogLevel MinimumLogLevel { get; set; } = LogLevel.Warning;
 }