+++ release_notes.md
@@ -2,11 +2,14 @@
 
 #### v1.0.1 2026-07-??
 
+* Add `loglevel` setting to [app] section (default is Warning)
+* Add `validate` interactive command to check `uptime.conf`
 * Show Windows permissions reminder when running service as user
-* Added warning about provider-wildcard overhead and filtering
+* README warning about provider-wildcard overhead and filtering
 * Bugfix - provider process-filter
 * Bugfix - include/exclude specifier parsing
-* Refactor - unified namespace now that it's a single codebase
+* Refactor - config parsing to support validation (including startup)
+* Refactor - unified namespace (now that it's a single codebase)
 
 #### v1.0.0 2026-07-19